requirements.txt
appnope==0.1.2
atlassian-python-api==3.8.0
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.7
defusedxml==0.7.1
Deprecated==1.2.12
idna==2.10
importlib-metadata==4.0.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
keyring==23.0.1
oauthlib==3.1.0
parso==0.8.2
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
python-dotenv==0.17.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.1
